+++ src/kspdg/__init__.py
@@ -57,7 +57,7 @@ try:
 except ModuleNotFoundError:
     print(f"Module {__evaluate_path} not found.")
 
-# Condition import of LG3 environments on the presence of juliacall dependency
+# Condition import of LG3, LG4 environments on the presence of juliacall dependency
 # Therefore, the kspdg library should be usable without the adv_bots optional 
 # dependency
 if find_spec('juliacall') is not None:
@@ -70,9 +70,24 @@ if find_spec('juliacall') is not None:
     LBG1_LG3_I1_V1 = getattr(__lg3_envs_module, 'LBG1_LG3_I1_Env')
     LBG1_LG3_I2_V1 = getattr(__lg3_envs_module, 'LBG1_LG3_I2_Env')
 
+    # import obfuscated LBG1-LG4 environments
+    __lg4_envs_path = __get_mod_str("kspdg_envs.lbg1.lg4_envs")
+    try:
+        __lg4_envs_module = importlib.import_module(__lg4_envs_path)
+    except ModuleNotFoundError:
+        print(f"Module {__lg4_envs_path} not found.")
+    LBG1_LG4_I1_V1 = getattr(__lg4_envs_module, 'LBG1_LG4_I1_Env')
+    LBG1_LG4_I2_V1 = getattr(__lg4_envs_module, 'LBG1_LG4_I2_Env')
+
 else:
     LBG1_LG3_I1_V1 = LBG1_LG3_I2_V1 = lambda *args, **kwargs: (
         "Unmet dependency juliacall for using LBG1_LG3 environments.\n" +
+        "Please install kspdg[adv_bots] or kspdg[full] to use these.\n" +
+        "Refere to README for further instructions."
+    )
+
+    LBG1_LG4_I1_V1 = LBG1_LG4_I2_V1 = lambda *args, **kwargs: (
+        "Unmet dependency juliacall for using LBG1_LG4 environments.\n" +
         "Please install kspdg[adv_bots] or kspdg[full] to use these.\n" +
         "Refere to README for further instructions."
     )