+++ src/kspdg/private_src/python3_12/Linux_x86_64/pyarmor_runtime_000000/__init__.py
@@ -1,2 +1,2 @@
-# Pyarmor 8.5.11 (trial), 000000, 2024-12-02T12:23:51.412694
+# Pyarmor 8.5.11 (trial), 000000, 2024-12-03T11:12:54.966912
 from .pyarmor_runtime import __pyarmor__