+++ src/kspdg/private_src/python3_13/Linux_x86_64/pyarmor_runtime_000000/__init__.py
@@ -1,2 +1,2 @@
-# Pyarmor 9.0.6 (trial), 000000, 2024-12-02T12:23:47.875181
+# Pyarmor 9.0.6 (trial), 000000, 2024-12-03T11:12:52.007265
 from .pyarmor_runtime import __pyarmor__