+++ src/kspdg/private_src/python3_13/Windows_x86_64/pyarmor_runtime_000000/__init__.py
@@ -1,2 +1,2 @@
-# Pyarmor 9.0.6 (trial), 000000, 2024-12-02T12:23:47.548622
+# Pyarmor 9.0.6 (trial), 000000, 2024-12-03T11:12:51.763259
 from .pyarmor_runtime import __pyarmor__