+++ src/kspdg/private_src/python3_9/Linux_x86_64/pyarmor_runtime_000000/__init__.py
@@ -1,2 +1,2 @@
-# Pyarmor 8.5.11 (trial), 000000, 2024-12-02T12:23:53.959975
+# Pyarmor 8.5.11 (trial), 000000, 2024-12-03T11:12:57.854777
 from .pyarmor_runtime import __pyarmor__