+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] - XXXX.XX.XX
+
+### Added
+
++ LBG1_LG4 private-source environments and unit-tests
+
 ## [v0.9.2] - 2024.12.03
 
 ### Fixed